requirements.txt
aiocontextvars==0.2.2
aiohttp==3.6.2
aiohttp-cors==0.7.0
aniso8601==8.0.0
async-timeout==3.0.1
attrs==19.3.0
certifi==2019.9.11
chardet==3.0.4
Click==7.0
colorclass==2.2.0
colorlog==4.0.2
contextvars==2.4
croniter==0.3.30
faust==1.9.0
Flask==1.1.1
Flask-RESTful==0.3.7
idna==2.8
idna-ssl==1.1.0
immutables==0.11
itsdangerous==1.1.0
Jinja2==2.10.1
kafka-python==1.4.6
MarkupSafe==1.1.1
mode==4.1.3
multidict==4.5.2
mypy-extensions==0.4.3
nltk==3.4.5
numpy==1.17.3
oauthlib==3.1.0
opentracing==1.3.0
pandas==0.25.2
psycopg2==2.8.4
PySocks==1.7.1
python-dateutil==2.8.0
pytz==2019.3
requests==2.22.0
requests-oauthlib==1.2.0
robinhood-aiokafka==1.1.3
six==1.12.0
terminaltables==3.1.0
textblob==0.15.3
tweepy==3.8.0
tweet-preprocessor==0.5.0
typing-extensions==3.7.4.1
urllib3==1.25.6
venusian==1.2.0
Werkzeug==0.16.0
yarl==1.3.0
